# server/app/static/openapi.yaml
openapi: 3.1.0
info:
  title: QueryBot API
  version: 1.0.0
  description: |
    Natural-language analytics over uploaded datasets.

    ## Authentication
    Sign-in sets two httpOnly cookies, `access_token` and `refresh_token`. The
    browser attaches them automatically; no token is ever readable from
    JavaScript. When the access token expires the client calls
    `POST /auth/refresh` once and replays the original request.

    ## Authorization
    Datasets belong to the user who uploaded them. Every endpoint that reads a
    dataset verifies ownership and answers `404` — not `403` — for a dataset
    owned by someone else, so identifiers cannot be probed for existence.

    ## Rate limits
    `POST /langgraph/run` calls a paid model several times per question and is
    limited to 20 per hour per user. Responses carry `X-RateLimit-*` headers.

    ## Tracing
    Every response includes `X-Request-ID`. Supply your own to correlate a call
    across all four services.
  license:
    name: ISC
servers:
  - url: http://localhost:8080/api
    description: Docker Compose
  - url: http://localhost:5000/api
    description: Local development
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
    Success:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { type: object }
    Error:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullname: { type: string }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
        is_active: { type: boolean }
...
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found, or owned by another user.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests.
      headers:
        X-RateLimit-Remaining: { schema: { type: integer } }
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - accessCookie: []
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullname, email, password, confirm_password]
              properties:
                fullname: { type: string, minLength: 2, maxLength: 100 }
                email: { type: string, format: email }
                password:
                  type: string
                  minLength: 8
                  description: Needs upper and lower case, a digit and a symbol.
                confirm_password: { type: string }
      responses:
        '201':
          description: Created; sets the auth cookies.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Success' }
        '409':
          description: That email is already registered.
        '422':
          description: Validation failed.
  /auth/login:
    post:
      tags: [Authentication]
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Signed in; sets the auth cookies.
        '401':
          description: Invalid credentials.
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Issue a new access token from the refresh cookie
      responses:
        '200': { description: A new access cookie was set. }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/check:
    get:
      tags: [Authentication]
      summary: Return the signed-in user
      responses:
        '200':
          description: The current user.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /langgraph/run:
    post:
      tags: [Analysis]
      summary: Ask a question about a dataset
      description: |
        Streams Server-Sent Events. Each event is a JSON object keyed by the
        workflow node that produced it, letting the client show progress as the
        agent classifies the question, writes SQL, runs it, renders a chart and
        derives insights. The final event is `{"done": true, "conversationId": "..."}`.

        Omit `conversationId` to start a new conversation; supply it to ask a
        follow-up, and earlier turns are passed to the agent as context.

        Rate limited to 20 per hour per user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question, databaseUuid]
              properties:
                question: { type: string, maxLength: 2000 }
                databaseUuid: { type: string, format: uuid }
                conversationId: { type: [string, 'null'], format: uuid }
      responses:
        '200':
          description: An SSE stream.
          headers:
            X-Conversation-Id:
              schema: { type: string, format: uuid }
          content:
            text/event-stream:
              schema: { type: string }
        '400': { description: Missing question or dataset. }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /langgraph/query:
    post:
      tags: [Analysis]
      summary: Run SQL directly
      description: |
        Executes a statement against the caller's dataset without invoking the
        model, which is what makes the generated SQL editable and instantly
        re-runnable. Only a single read-only `SELECT` or `WITH` is accepted;
        `ATTACH`, `PRAGMA` and `load_extension` are rejected and the database is
        opened read-only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query, databaseUuid]
              properties:
                query: { type: string }
                databaseUuid: { type: string, format: uuid }
      responses:
        '200':
          description: Query results.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          columns: { type: array, items: { type: string } }
                          results: { type: array, items: { type: array } }
                          rowCount: { type: integer }
                          truncated: { type: boolean }
                          durationMs: { type: integer }
        '400': { description: Rejected or invalid SQL. }
        '404': { $ref: '#/components/responses/NotFound' }
...